+++ aioxmpp/muc/__init__.py
@@ -84,6 +84,29 @@ XSOs
 
 .. currentmodule:: aioxmpp.muc.xso
 
+Attributes added to existing XSOs
+---------------------------------
+
+.. attribute:: aioxmpp.Message.xep0045_muc
+
+    A :class:`GenericExt` object or :data:`None`.
+
+.. attribute:: aioxmpp.Message.xep0045_muc_user
+
+    A :class:`UserExt` object or :data:`None`.
+
+.. attribute:: aioxmpp.Presence.xep0045_muc
+
+    A :class:`GenericExt` object or :data:`None`.
+
+.. attribute:: aioxmpp.Presence.xep0045_muc_user
+
+    A :class:`UserExt` object or :data:`None`.
+
+.. attribute:: aioxmpp.Message.xep0249_direct_invite
+
+    A :class:`DirectInvite` object or :data:`None`.
+
 Generic namespace
 -----------------
 
@@ -125,6 +148,11 @@ Owner namespace
 .. autoclass:: OwnerQuery
 
 .. autoclass:: DestroyRequest
+
+:xep:`249` Direct Invitations
+-----------------------------
+
+.. autoclass:: DirectInvite
 
 """
 from .service import MUCClient, Occupant, Room, LeaveMode, RoomState  # NOQA